+++ .gitignore
@@ -1,7 +1,6 @@
 node_modules
 .DS_Store
 dist
-dist-ssr
 *.local
 count.txt
 .env
@@ -9,8 +8,6 @@ count.txt
 .nitro
 .tanstack
 .output
-.vinxi
-count.txt
 *.swp
-.netlify
 terraform.tfvars
+scratch
