+++ package.json
@@ -2,6 +2,9 @@
   "name": "jspsi",
   "version": "1.0.0",
   "description": "JavaScript implementation of private set intersection for record linkage using peer-to-peer connections.",
+  "license": "Apache-2.0",
+  "author": "Vincent Dorie <vdorie@gmail.com>",
+  "type": "module",
   "main": "index.tsx",
   "scripts": {
     "dev": "vite dev",
@@ -19,8 +22,6 @@
     "private set intersection",
     "record linkage"
   ],
-  "author": "",
-  "type": "module",
   "dependencies": {
     "@mantine/core": "^8.1.3",
     "@mantine/dropzone": "^8.2.1",
@@ -38,6 +39,7 @@
     "env-schema": "^6.0.1",
     "loglevel": "^1.9.2",
     "peerjs": "^1.5.5",
+    "psi-link": "file:psi-link/psi-link-1.0.0.tgz",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "ssh2-sftp-client": "^12.0.1",
