+++ psi-link/tsconfig.json
@@ -1,0 +1,23 @@
+{
+  "include": [
+    "src/**/*.ts",
+    "test/**/*.ts",
+    "eslint.config.js",
+    "rollup.config.ts",
+    "vitest.config.ts"
+  ],
+  "compilerOptions": {
+    "target": "es2022",
+    "module": "esnext",
+    "moduleResolution": "bundler",
+    "verbatimModuleSyntax": true,
+
+    "strict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noFallthroughCasesInSwitch": true,
+    "noUncheckedSideEffectImports": true,
+    "strictNullChecks": true,
+    "forceConsistentCasingInFileNames": true,
+  }
+}
