+++ psi-link/vitest.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    include: [
+      "test/**/*.{test,spec}.?(c|m)[jt]s?(x)"
+    ],
+  }
+});
